# docs/openapi.yaml
openapi: 3.0.0
info:
  title: superglue AI API
  version: 1.0.0
  description: API for running superglue AI tools
  contact:
    name: superglue AI Support
    email: stefan@superglue.ai
servers:
  - url: https://api.superglue.ai/v1
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /end-users/{endUserId}:
    get:
      summary: Get end user details
      operationId: getEndUser
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Get end user details including credential connection status for all systems.
      parameters:
        - name: endUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: End user details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/EndUser"
                      - type: object
                        properties:
                          credentials:
                            type: array
                            description: Credential status for each system
                            items:
                              type: object
                              properties:
                                systemId:
                                  type: string
                                systemName:
                                  type: string
                                hasCredentials:
                                  type: boolean
        "404":
          description: End user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete end user
      operationId: deleteEndUser
      tags:
        - End Users (Enterprise)
      description: |
        **Enterprise feature.**
        
        Delete an end user and all their credentials. This action cannot be undone.
      parameters:
        - name: endUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: End user deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "404":
          description: End user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /hooks/{toolId}:
    post:
      summary: Handle incoming webhook
      operationId: triggerWebhook
      tags:
        - Webhooks
      description: |
        **Enterprise feature.**

        Trigger a tool execution from an external webhook (Stripe, GitHub, etc.).

        The request body becomes the tool's input payload.
        Returns 202 Accepted immediately and executes the tool asynchronously.
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
          description: ID of the tool to trigger
          example: handle-stripe-webhook
      requestBody:
        required: false
        description: Webhook payload from the external service. Becomes the tool's input.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                id: evt_1234
                type: customer.created
                data:
                  object:
                    id: cus_abc123
                    email: user@example.com
      responses:
        "202":
          description: Webhook accepted, tool execution started
          headers:
            X-Trace-Id:
              schema:
                type: string
              description: Trace ID for this request
          content:
            application/json:
              schema:
                type: object
                required:
                  - runId
                  - status
                  - toolId
                properties:
                  runId:
                    type: string
                    description: ID of the created run
                    example: 7f3e9c1a-2b4d-4e8f-9a3b-1c5d7e9f2a4b
                  status:
                    type: string
                    enum: [accepted]
                    example: accepted
                  toolId:
                    type: string
                    description: ID of the triggered tool
                    example: handle-stripe-webhook
        "400":
          description: Tool is archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: Static API Key
      description: |
        Static API key authentication using Bearer token scheme.
        Include your API key in the Authorization header: `Authorization: Bearer YOUR_API_KEY`

        Alternatively, you can use the `token` query parameter to authenticate.

        API keys can be generated in your superglue dashboard.
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              example: Invalid input parameters
    EndUser:
      type: object
      description: |
        An end user who can connect their own credentials to multi-tenancy systems.
      required:
        - id
        - externalId
      properties:
        id:
          type: string
          format: uuid
          description: Internal superglue user ID
          example: 550e8400-e29b-41d4-a716-446655440000
        externalId:
          type: string
          description: Your application's user ID
          example: user-123
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        metadata:
          type: object
          description: Custom metadata for this user
          additionalProperties: true
        allowedSystems:
          type: array
          description: System IDs this user can access, or ["*"] for all systems
          items:
            type: string
          example: ["*"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time